separate WrappedCandidatePreamble into a separate file for clarity, raised speaking window
